refactor(types): centralize CalendarEvent cluster into types/calendar.ts
- New file src/app/types/calendar.ts is now the canonical SSOT for the
  student calendar domain: CalendarEvent, HeatmapEntry, CalendarTask,
  CalendarData (all DB-backed shapes consumed by GET /calendar/data).
- useCalendarEvents.ts imports these from the new module and re-exports
  them for backward compatibility with existing call sites.
- 15 real consumers updated to import directly from @/app/types/calendar
  (hooks useFinalsWeek/useHeatmap, calendar/* components, test files).
- Two view files (KnowledgeHeatmapView, MasteryDashboardView) had local
  file-scoped interfaces named CalendarEvent used ONLY for hardcoded
  mock-data arrays. Their shape is fundamentally different from the DB
  entity ({day: number, category, cards?, ...}). Renamed to
  MockCalendarDayEvent with a JSDoc clarifying they are not the canonical
  type. No behavior change.
- Summary type: no edits needed. It is already SSOT in summariesApi.ts,
  re-exported by platform.ts. student.ts only contains StudySummary (a
  distinct type). Audit claim of 3 duplicates was stale.
- Bonus: fix pre-existing TS2304 in types/platform.ts line 231
  (ContentHierarchy.summaries: Summary[] could not resolve because the
  file only re-exported Summary without importing it locally). Changed
  to `import type { Summary } + export type { Summary }` so the type is
  both in local scope and re-exported. tsc delta: -1 error vs baseline.

Validation: tsc --noEmit --ignoreDeprecations 6.0 run before and after.
Total errors 22212 (clean main) -> 22211 (with changes). All remaining
errors are pre-existing "Cannot find module 'react/vitest/..." from the
absent node_modules in the worktree; none are attributable to this
consolidation.

Scope note: Tool-sandboxed files (agent memory under .claude/) could not
be edited. Session notes are captured here and in the team report.
