+++ src/app/hooks/useCalendarEvents.ts
@@ -14,42 +14,17 @@ import { useQuery } from '@tanstack/react-query';
 import { useMemo } from 'react';
 import { format } from 'date-fns';
 import { apiCall } from '@/app/lib/api';
+import type {
+  CalendarEvent,
+  HeatmapEntry,
+  CalendarTask,
+  CalendarData,
+} from '@/app/types/calendar';
 
-// ── Types ───────────────────────────────────────────────────
-
-export interface CalendarEvent {
-  id: string;
-  student_id: string;
-  course_id: string;
-  institution_id: string;
-  title: string;
-  date: string;          // ISO date string YYYY-MM-DD
-  time: string | null;
-  location: string | null;
-  is_final: boolean;
-  exam_type: string;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface HeatmapEntry {
-  date: string;          // ISO date string YYYY-MM-DD
-  minutes: number;       // Total study minutes for this day
-}
-
-export interface CalendarTask {
-  id: string;
-  student_id: string;
-  title: string;
-  scheduled_date: string;
-  status: string;
-}
-
-export interface CalendarData {
-  events: CalendarEvent[];
-  heatmap: HeatmapEntry[];
-  tasks: CalendarTask[];
-}
+// ── Types (canonical shapes live in @/app/types/calendar) ───
+// Re-exported here for backward compatibility with existing
+// consumers that import from '@/app/hooks/useCalendarEvents'.
+export type { CalendarEvent, HeatmapEntry, CalendarTask, CalendarData };
 
 // ── Query Key ───────────────────────────────────────────────
